@@ -1,7 +1,7 @@
 {
   "name": "violentmonkey",
   "description": "Violentmonkey",
-  "version": "2.20.0",
+  "version": "2.29.0",
   "scripts": {
     "prepare": "husky install",
     "dev": "gulp dev",
@@ -35,9 +35,9 @@
     "@types/jest": "^29.5.3",
     "@typescript-eslint/eslint-plugin": "^6.4.0",
     "@typescript-eslint/parser": "^6.4.0",
-    "@violentmonkey/types": "0.1.5",
-    "@vue/compiler-sfc": "^3.3.4",
-    "@vue/eslint-config-typescript": "^11.0.3",
+    "@violentmonkey/types": "0.2.0",
+    "@vue/compiler-sfc": "^3.4.38",
+    "@vue/eslint-config-typescript": "^13.0.0",
     "amo-upload": "^0.5.5",
     "autoprefixer": "^10.4.15",
     "babel-jest": "^29.6.2",
@@ -47,20 +47,20 @@
     "cross-env": "^7.0.3",
     "cross-spawn": "^7.0.3",
     "css-loader": "^6.8.1",
-    "css-minimizer-webpack-plugin": "^5.0.1",
+    "css-minimizer-webpack-plugin": "^7.0.0",
     "del": "^6.1.1",
-    "eslint": "^8.47.0",
+    "eslint": "^8.57.1",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-jest": "^27.2.3",
     "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint-plugin-vue": "^9.27.0",
     "fancy-log": "^2.0.0",
     "gulp": "^4.0.2",
     "gulp-plumber": "^1.1.0",
-    "html-webpack-plugin": "^5.5.3",
+    "html-webpack-plugin": "^5.6.0",
     "husky": "^8.0.1",
-    "jest": "^29.6.2",
-    "jest-environment-jsdom": "^29.0.3",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
     "mini-css-extract-plugin": "^2.7.6",
@@ -81,12 +81,11 @@
     "through2": "^4.0.2",
     "typescript": "^5.1.6",
     "vinyl": "^2.2.1",
-    "vue-loader": "^17.2.2",
-    "vue-tsc": "^1.8.8",
-    "webpack": "^5.88.2",
-    "webpack-bundle-analyzer": "^4.9.0",
-    "webpack-cli": "^5.1.4",
-    "wrapper-webpack-plugin": "2.2.2"
+    "vue-loader": "^17.4.2",
+    "vue-tsc": "^2.0.29",
+    "webpack": "^5.95.0",
+    "webpack-bundle-analyzer": "^4.10.2",
+    "webpack-cli": "^5.1.4"
   },
   "author": "Gerald <i@gerald.top>",
   "repository": {
@@ -103,7 +102,7 @@
     "@zip.js/zip.js": "2.4.4",
     "codemirror": "^5.65.10",
     "tldts": "^6.0.14",
-    "vue": "^3.3.4",
+    "vue": "^3.4.38",
     "vueleton": "^2.0.2"
   },
   "engines": {
