@@ -7,10 +7,14 @@ import { popupTabs } from './popup-tracker';
 import storage, { S_CACHE } from './storage';
 import { forEachTab, getTabUrl, injectableRe, openDashboard, tabsOnRemoved, tabsOnUpdated } from './tabs';
 import { testBlacklist } from './tester';
-import { FIREFOX } from './ua';
+import { FIREFOX, ua } from './ua';
 
-/** We don't set 19px because FF and Vivaldi scale it down to 16px instead of our own crisp 16px */
-const SIZES = [16, 32];
+/** 1x + HiDPI 1.5x, 2x */
+const SIZES = !FIREFOX
+  ? [16, 32]
+  : ua.mobile
+    ? [32, 38, 48] // 1x, 1.5x, 2x
+    : [16, 32, 48, 64]; // 16+32: toolbar, 32+48+64: extensions panel
 /** Caching own icon to improve dashboard loading speed, as well as browserAction API
  * (e.g. Chrome wastes 40ms in our extension's process to read 4 icons for every tab). */
 const iconCache = {};
@@ -245,6 +249,8 @@ export function handleHotkeyOrMenu(id, tab) {
     openDashboard('');
   } else if (id === 'newScript') {
     commands.OpenEditor();
+  } else if (id === 'toggleInjection') {
+    setOption(IS_APPLIED, !isApplied);
   } else if (id === 'updateScripts') {
     commands.CheckUpdate();
   } else if (id === 'updateScriptsInTab') {
@@ -260,7 +266,8 @@ async function loadIcon(url) {
   const isOwn = url.startsWith(ICON_PREFIX);
   img.src = isOwn ? url.slice(extensionOrigin.length) // must be a relative path in Firefox Android
     : url.startsWith('data:') ? url
-      : makeDataUri(url[0] === 'i' ? url : await loadStorageCache(url));
+      : makeDataUri(url[0] === 'i' ? url : await loadStorageCache(url))
+        || url;
   await new Promise((resolve) => {
     img.onload = resolve;
     img.onerror = resolve;
