@@ -1,12 +1,11 @@
 import { browserWindows, request, noop, i18n, getUniqId } from '@/common';
+import { FILE_GLOB_ALL } from '@/common/consts';
 import cache from './cache';
 import { addPublicCommands, commands } from './init';
 import { getOption } from './options';
-import { parseMeta, isUserScript } from './script';
+import { parseMeta, matchUserScript } from './script';
 import { fileSchemeRequestable, getTabUrl, NEWTAB_URL_RE, tabsOnUpdated } from './tabs';
 import { FIREFOX } from './ua';
-
-const CONFIRM_URL_BASE = `${extensionRoot}confirm/index.html#`;
 
 addPublicCommands({
   async CheckInstallerTab(tabId, src) {
@@ -14,43 +13,53 @@ addPublicCommands({
       && await browser.tabs.get(tabId).catch(noop);
     return tab && getTabUrl(tab).startsWith(CONFIRM_URL_BASE);
   },
-  async ConfirmInstall({ code, from, url, fs }, { tab = {} }) {
-    if (!fs) {
-      if (!code) code = (await request(url)).data;
-      // TODO: display the error in UI
-      if (!isUserScript(code)) {
-        throw `${i18n('msgInvalidScript')}\n\n${
-          code.trim().split(/[\r\n]+\s*/, 9/*max lines*/).join('\n')
-            .slice(0, 500/*max overall length*/)
-        }...`;
-      }
-      cache.put(url, code, 3000);
-    }
-    const confirmKey = getUniqId();
-    const { active, id: tabId, incognito } = tab;
-    // Not testing tab.pendingUrl because it will be always equal to `url`
-    const canReplaceCurTab = (!incognito || IS_FIREFOX) && (
-      url === from
-      || cache.has(`autoclose:${tabId}`)
-      || NEWTAB_URL_RE.test(from));
-    /** @namespace VM.ConfirmCache */
-    cache.put(`confirm-${confirmKey}`, { incognito, url, from, tabId, fs, ff: FIREFOX });
-    const confirmUrl = CONFIRM_URL_BASE + confirmKey;
-    const { [kWindowId]: windowId } = canReplaceCurTab
-      ? await browser.tabs.update(tabId, { url: confirmUrl })
-      : await commands.TabOpen({ url: confirmUrl, active: !!active }, { tab });
-    if (active && windowId !== tab[kWindowId]) {
-      await browserWindows?.update(windowId, { focused: true });
-    }
-  },
+  ConfirmInstall: confirmInstall,
 });
 
+async function confirmInstall({ code, from, url, fs, parsed }, { tab = {} }) {
+  if (!fs) {
+    code ??= parsed
+      ? request(url).then(r => r.data) // cache the Promise and start fetching now
+      : (await request(url)).data;
+    // TODO: display the error in UI
+    if (!parsed && !matchUserScript(code)) {
+      throw `${i18n('msgInvalidScript')}\n\n${
+        code.trim().split(/[\r\n]+\s*/, 9/*max lines*/).join('\n')
+          .slice(0, 500/*max overall length*/)
+      }...`;
+    }
+    cache.put(url, code, 3000);
+  }
+  const confirmKey = getUniqId();
+  const { active, id: tabId, incognito } = tab;
+  // Not testing tab.pendingUrl because it will be always equal to `url`
+  const canReplaceCurTab = (!incognito || IS_FIREFOX) && (
+    url === from
+    || cache.has(`autoclose:${tabId}`)
+    || NEWTAB_URL_RE.test(from));
+  /** @namespace VM.ConfirmCache */
+  cache.put(`confirm-${confirmKey}`, { incognito, url, from, tabId, fs, ff: FIREFOX });
+  const confirmUrl = CONFIRM_URL_BASE + confirmKey;
+  const { [kWindowId]: windowId } = canReplaceCurTab
+  // The tab may have been closed already, in which case we'll open a new tab
+  && await browser.tabs.update(tabId, { url: confirmUrl }).catch(noop)
+  || await commands.TabOpen({ url: confirmUrl, active: !!active }, { tab });
+  if (active && windowId !== tab[kWindowId]) {
+    await browserWindows?.update(windowId, { focused: true });
+  }
+}
+
+const CONFIRM_URL_BASE = `${extensionRoot}confirm/index.html#`;
 const whitelistRe = re`/^https:\/\/(
   (greas|sleaz)yfork\.org\/scripts\/[^/]*\/code|
+  update\.(greas|sleaz)yfork\.org\/scripts|
   openuserjs\.org\/install\/[^/]*|
   github\.com\/[^/]*\/[^/]*\/(
     raw\/[^/]*|
-    releases\/download\/[^/]*
+    releases\/(
+      download\/[^/]* |
+      latest\/download
+    )
   )|
   raw\.githubusercontent\.com(\/[^/]*){3}|
   gist\.github\.com\/.*?
@@ -72,15 +81,25 @@ const maybeRedirectVirtualUrlFF = virtualUrlRe && ((tabId, src) => {
   }
 });
 
-async function maybeInstallUserJs(tabId, url) {
+async function maybeInstallUserJs(tabId, url, isWhitelisted) {
   // Getting the tab now before it navigated
   const tab = tabId >= 0 && await browser.tabs.get(tabId) || {};
-  const { data: code } = await request(url).catch(noop) || {};
-  if (code && parseMeta(code).name) {
-    commands.ConfirmInstall({ code, url, from: tab.url }, { tab });
+  const { data: code } = !isWhitelisted && await request(url).catch(noop) || {};
+  if (isWhitelisted || code && parseMeta(code).name) {
+    confirmInstall({ code, url, from: tab.url, parsed: true }, { tab });
   } else {
     cache.put(`bypass:${url}`, true, 10e3);
-    if (tabId >= 0) browser.tabs.update(tabId, { url });
+    const error = `${VIOLENTMONKEY} installer skipped ${url}.
+Either not a userscript or the metablock comment is malformed:
+${code?.length > 1e6 ? code.slice(0, 1e6) + '...' : code}`;
+    if (tabId < 0) {
+      console.warn(error);
+    } else {
+      browser.tabs.executeScript(tabId, {
+        code: `console.warn(${JSON.stringify(error)})`,
+      });
+      browser.tabs.update(tabId, { url });
+    }
   }
 }
 
@@ -106,7 +125,7 @@ browser.tabs.onCreated.addListener((tab) => {
   }
   if (isUserJS && isFile && !fileSchemeRequestable && !IS_FIREFOX
   && getOption('helpForLocalFile')) {
-    commands.ConfirmInstall({ url, fs: true }, { tab });
+    confirmInstall({ url, fs: true }, { tab });
   }
 });
 
@@ -119,9 +138,10 @@ browser.webRequest.onBeforeRequest.addListener((req) => {
   if (url.startsWith(extensionRoot)) {
     return { redirectUrl: resolveVirtualUrl(url) };
   }
+  let isWhitelisted;
   if (!cache.has(`bypass:${url}`)
-  && (!blacklistRe.test(url) || whitelistRe.test(url))) {
-    maybeInstallUserJs(tabId, url);
+  && ((isWhitelisted = whitelistRe.test(url)) || !blacklistRe.test(url))) {
+    maybeInstallUserJs(tabId, url, isWhitelisted);
     return IS_FIREFOX
       ? { cancel: true } // for sites with strict CSP in FF
       : { redirectUrl: 'javascript:void 0' }; // eslint-disable-line no-script-url
@@ -133,8 +153,8 @@ browser.webRequest.onBeforeRequest.addListener((req) => {
     // 3. Firefox: onBeforeRequest does not work with file:// or moz-extension://
     '*://*/*.user.js',
     '*://*/*.user.js?*',
-    'file://*/*.user.js',
-    'file://*/*.user.js?*',
+    `${FILE_GLOB_ALL}.user.js`,
+    `${FILE_GLOB_ALL}.user.js?*`,
     `${extensionRoot}*.user.js`,
   ],
   types: ['main_frame'],
