@@ -1,6 +1,6 @@
 <template>
   <label class="setting-check">
-    <input type="checkbox" v-model="value" :disabled="disabled">
+    <input type="checkbox" v-model="value" :disabled>
     <slot>
       <span v-text="label" />
     </slot>
@@ -44,5 +44,13 @@ onBeforeUnmount(revoke);
 <style>
 .setting-check {
   display: inline-flex;
+  flex-wrap: wrap;
+  white-space: pre-wrap; /* preserving spaces in html */
+  > :nth-child(2) {
+    flex: 1 1 min-content; /* wrapping inside the label so it stays in the same row as [x] */
+  }
+  .vl-dropdown-menu & {
+    width: max-content; /* CSS bug(?) workaround for absolutely positioned containers */
+  }
 }
 </style>
