@@ -2,8 +2,8 @@ const handlers = createNullObj();
 export const addHandlers = obj => assign(handlers, obj);
 export const callbacks = createNullObj();
 /**
- * @property {VMScriptGMInfoPlatform} ua
- * @property {VMBridgePostFunc} post
+ * @mixes VMInjection.Info
+ * @property {VMBridgePostFunc} post - synchronous
  * @property {VMBridgeMode} mode
  */
 const bridge = {
@@ -12,19 +12,18 @@ const bridge = {
     const fn = handlers[cmd];
     if (fn) node::fn(data);
   },
-  send(cmd, data, node) {
+  /** @return {Promise} asynchronous */
+  promise(cmd, data, node) {
     let cb;
     let res;
-    try {
-      res = new UnsafePromise(resolve => {
-        cb = resolve;
-      });
-    } catch (e) {
-      // Unavoidable since vault's Promise can't be used after the iframe is removed
-    }
+    res = new SafePromise(resolve => {
+      cb = resolve;
+    });
+    if (IS_FIREFOX) setPrototypeOf(res, SafePromiseConstructor);
     postWithCallback(cmd, data, node, cb);
     return res;
   },
+  /** @return {?} synchronous */
   call: postWithCallback,
 };
 
