@@ -1,7 +1,7 @@
 import bridge from './bridge';
 import { GM4_ALIAS, GM_API, GM_API_CTX, GM_API_CTX_GM4ASYNC } from './gm-api';
 import { makeGlobalWrapper } from './gm-global-wrapper';
-import { makeComponentUtils, safeAssign } from './util';
+import { makeComponentUtils, safeCopy, thisObjectProto } from './util';
 
 /** @type {(keyof VMInjection.Script)[]} */
 const COPY_SCRIPT_PROPS = [
@@ -9,6 +9,11 @@ const COPY_SCRIPT_PROPS = [
   'id',
 ];
 const componentUtils = makeComponentUtils();
+const kResources = 'resources';
+const getUA = () => bridge.call('UA');
+const getUAHints = hints => bridge.promise('UAH', hints);
+const getUAData = () => bridge.uad
+  && setOwnProp(bridge.call('UAD'), 'getHighEntropyValues', getUAHints);
 const sendTabClose = () => bridge.post('TabClose');
 const sendTabFocus = () => bridge.post('TabFocus');
 
@@ -21,22 +26,18 @@ export function makeGmApiWrapper(script) {
   // Reference: http://wiki.greasespot.net/Greasemonkey_Manual:API
   const { meta } = script;
   const { grant } = meta;
-  const resources = setPrototypeOf(meta.resources, null);
+  const resources = setPrototypeOf(meta[kResources], null);
   /** @type {GMContext} */
   const context = safePickInto({
-    resources,
+    [kResources]: resources,
     resCache: createNullObj(),
     async: false,
   }, script, COPY_SCRIPT_PROPS);
-  const gmInfo = makeGmInfo(script.gmi, meta, resources);
-  const gm4 = {
-    __proto__: null,
-    info: gmInfo,
-  };
+  const gmInfo = script.gmi;
+  const gm4 = createNullObj();
   const gm = {
     __proto__: null,
     GM: gm4,
-    GM_info: gmInfo,
     unsafeWindow: global,
   };
   let contextAsync;
@@ -46,6 +47,7 @@ export function makeGmApiWrapper(script) {
     numGrants = 0;
   }
   assign(gm, componentUtils);
+  defineGmInfoProps(makeGmInfo, 'get');
   for (let name of grant) {
     let fn, fnGm4, gmName, gm4name;
     if (name::slice(0, 3) === 'GM.' && (gm4name = name::slice(3)) && (fnGm4 = GM4_ALIAS[gm4name])
@@ -73,22 +75,27 @@ export function makeGmApiWrapper(script) {
     gm.c = gm;
   }
   return { gm, wrapper };
-}
 
-function makeGmInfo(gmInfo, meta, resources) {
-  const resourcesArr = objectKeys(resources);
-  resourcesArr::forEach((name, i) => {
-    resourcesArr[i] = { name, url: resources[name] };
-  });
-  // No __proto__:null because these are standard objects for userscripts
-  meta.resources = resourcesArr;
-  safeAssign(gmInfo, bridge.gmi);
-  return safeAssign(gmInfo, {
-    [INJECT_INTO]: bridge.mode,
-    platform: safeAssign({}, bridge.ua),
-    script: meta,
-    scriptHandler: VIOLENTMONKEY,
-    version: process.env.VM_VER,
-  });
+  function defineGmInfoProps(value, getter) {
+    setOwnProp(gm, 'GM_info', value, true, getter);
+    setOwnProp(gm4, 'info', value, true, getter);
+  }
+  function makeGmInfo() {
+    setPrototypeOf(gmInfo, null); // enable safe direct assignment
+    const arr = meta[kResources] = objectKeys(resources);
+    for (let i = 0, name; i < arr.length; i++) {
+      name = arr[i];
+      arr[i] = { name, url: resources[name] };
+    }
+    assign(gmInfo, bridge.gmi);
+    gmInfo[INJECT_INTO] = bridge.mode;
+    gmInfo.platform = safeCopy(bridge.ua);
+    gmInfo.script = meta;
+    gmInfo.scriptHandler = VIOLENTMONKEY;
+    gmInfo.version = process.env.VM_VER;
+    setOwnProp(gmInfo, 'userAgent', getUA, true, 'get');
+    setOwnProp(gmInfo, 'userAgentData', getUAData, true, 'get');
+    defineGmInfoProps(gmInfo);
+    return setPrototypeOf(gmInfo, thisObjectProto); // return as a standard Object
+  }
 }
-
