@@ -113,9 +113,12 @@ export const makeComponentUtils = () => {
   };
 };
 
-export const safeAssign = (dst, src) => {
-  for (const key of objectKeys(src)) {
-    setOwnProp(dst, key, src[key]);
-  }
-  return dst;
-};
+/** Safe Object prototype.
+ * We can't use `Object` from safe-globals as it belongs to another frame.
+ * We can't use `global.Object` as it may have been spoofed. */
+export const thisObjectProto = getPrototypeOf({});
+
+export const safeCopy = src => setPrototypeOf(
+  assign(createNullObj(), src),
+  thisObjectProto
+);
