@@ -1,61 +1,9 @@
 <template>
-  <div class="tab-settings" :data-show-advanced="settings.showAdvanced">
+  <div ref="$el" class="tab-settings" :data-show-advanced="settings.showAdvanced">
     <h1 v-text="i18n('labelSettings')"></h1>
     <section class="mb-1c">
       <h3 v-text="i18n('optionPopup')"/>
-      <div>
-        <setting-check name="autoReload" :label="i18n('labelAutoReloadCurrentTab')" />
-      </div>
-      <div class="ml-2c flex flex-col">
-        <!-- ml-2c indents children after the first one for visual grouping -->
-        <setting-check name="editorWindow" class="mr-2" ref="EW">
-          <tooltip :content="editorWindowHint" :disabled="!editorWindowHint">
-            <span v-text="i18n('optionEditorWindow')"></span>
-          </tooltip>
-        </setting-check>
-        <setting-check name="editorWindowSimple" :label="i18n('optionEditorWindowSimple')"
-                       v-show="$refs.EW?.value"/>
-      </div>
-      <div class="ml-2c">
-        <label>
-          <locale-group i18n-key="labelPopupSort">
-            <select v-for="opt in ['filtersPopup.sort']" v-model="settings[opt]" :key="opt">
-              <option v-for="(title, value) in items[opt].enum" :key="`${opt}:${value}`"
-                      :value="value" v-text="title" />
-            </select>
-          </locale-group>
-        </label>
-        <setting-check name="filtersPopup.groupRunAt" :label="i18n('optionPopupGroupRunAt')"
-                       v-show="settings['filtersPopup.sort'] === 'exec'" />
-        <label>
-          <select v-for="opt in ['filtersPopup.hideDisabled']" v-model="settings[opt]" :key="opt">
-            <option v-for="(title, value) in items[opt].enum" :key="`${opt}:${value}`"
-                    :value="value" v-text="title" />
-          </select>
-        </label>
-        <setting-check name="filtersPopup.enabledFirst" :label="i18n('optionPopupEnabledFirst')"
-                       v-show="!settings['filtersPopup.hideDisabled']" />
-      </div>
-      <div>
-        <label>
-          <span v-text="i18n('labelBadge')"></span>
-          <select v-for="opt in ['showBadge']" v-model="settings[opt]" :key="opt">
-            <option v-for="(title, value) in items[opt].enum" :key="`${opt}:${value}`"
-                    :value="value" v-text="title" />
-          </select>
-        </label>
-      </div>
-      <div>
-        <label>
-          <span v-text="i18n('labelBadgeColors')"/>
-          <tooltip v-for="(title, name) in items.badgeColor.enum" :key="`bc:${name}`"
-                   :content="title">
-            <input type="color" v-model="settings[name]">
-          </tooltip>
-          <button v-text="i18n('buttonReset')" v-show="isCustomBadgeColor" class="ml-1"
-                  @click="onResetBadgeColors"/>
-        </label>
-      </div>
+      <settings-popup/>
     </section>
     <section class="mb-1c">
       <h3 v-text="i18n('optionUpdate')"/>
@@ -65,7 +13,8 @@
             <input v-model="settings.autoUpdate" type="number" min=0 max=365 step=1/>
           </locale-group>
         </label>
-        <setting-check name="updateEnabledScriptsOnly" :label="i18n('labelEnabledScriptsOnly')" />
+        <setting-check :name="kUpdateEnabledScriptsOnly"
+                       :label="i18n('labelEnabledScriptsOnly')" />
       </div>
       <div class="ml-2c flex flex-col">
         <setting-check name="notifyUpdates" :label="i18n('labelNotifyUpdates')" />
@@ -74,14 +23,14 @@
       </div>
     </section>
     <section class="mb-2c">
-      <h3 v-text="i18n('labelBackupMaintenance')" />
+      <h3 v-text="i18n('labelBackupMaintenance')" :class="{bright: store.isEmpty === 1}"/>
       <vm-import></vm-import>
       <vm-export></vm-export>
       <hr>
       <vm-maintenance/>
     </section>
     <vm-sync></vm-sync>
-    <details v-for="(obj, key) in {showAdvanced: settings}" :key="key" :open="obj[key]">
+    <details v-for="(obj, key) in {showAdvanced: settings}" :key :open="obj[key]">
       <summary @click.prevent="obj[key] = !obj[key]">
         <!-- eslint-disable-next-line vue/no-v-text-v-html-on-component -->
         <component v-text="i18n('labelAdvanced')" class="inline-block"
@@ -93,8 +42,8 @@
           <label>
             <locale-group i18n-key="optionUiTheme">
               <select v-for="opt in ['uiTheme']" v-model="settings[opt]" :key="opt">
-                <option v-for="(title, value) in items[opt].enum" :key="value"
-                        :value="value" v-text="title" />
+                <option v-for="(title, value) in items[opt]" :key="value"
+                        :value v-text="title" />
               </select>
             </locale-group>
           </label>
@@ -103,9 +52,9 @@
           <label>
             <span v-text="i18n('labelInjectionMode')"></span>
             <select v-for="opt in ['defaultInjectInto']" v-model="settings[opt]" :key="opt">
-              <option v-for="(_, mode) in items[opt].enum" :key="mode" v-text="mode" />
+              <option v-for="(_, mode) in items[opt]" :key="mode" v-text="mode" />
             </select>
-            <a class="ml-1" href="https://violentmonkey.github.io/posts/inject-into-context/" target="_blank" rel="noopener noreferrer" v-text="i18n('learnInjectionMode')"></a>
+            <a class="ml-1" :href="VM_HOME + 'posts/inject-into-context/'" v-bind="EXTERNAL_LINK_PROPS" v-text="i18n('learnInjectionMode')"/>
           </label>
           <tooltip :content="i18n('labelXhrInjectHint')">
             <setting-check name="xhrInject">
@@ -123,12 +72,12 @@
             </tooltip>
           </label>
         </div>
-        <div>
-          <locale-group i18n-key="labelExposeStatus" class="flex flex-col">
+        <div class="flex flex-col">
+          <locale-group i18n-key="labelExposeStatus">
             <setting-check v-for="([key, host]) in expose" :key="host"
-                           :name="`expose.${key}`" class="ml-2 mr-1c valign-tb">
+                           :name="`expose.${key}`" class="ml-2 mr-1c">
               <span v-text="host" />
-              <a :href="`https://${host}`" target="_blank" rel="noopener noreferrer">&nearr;</a>
+              <a :href="`https://${host}`" v-bind="EXTERNAL_LINK_PROPS">&nearr;</a>
             </setting-check>
           </locale-group>
         </div>
@@ -145,7 +94,7 @@
                      :key="i" :is="i % 2 ? 'code' : 'span'"
           /> <vm-date-info/><!--DANGER! Using the same line to preserve the space-->
         </p>
-        <setting-text name="scriptTemplate" has-reset/>
+        <setting-text :name="kScriptTemplate" has-reset/>
       </section>
 
       <vm-blacklist />
@@ -160,19 +109,35 @@
 </template>
 
 <script>
-import { reactive } from 'vue';
-import Tooltip from 'vueleton/lib/tooltip';
-import { debounce, i18n } from '@/common';
-import { KNOWN_INJECT_INTO } from '@/common/consts';
-import SettingCheck from '@/common/ui/setting-check';
-import { forEachEntry, mapEntry } from '@/common/object';
+import { i18n } from '@/common';
+import { KNOWN_INJECT_INTO, VM_HOME } from '@/common/consts';
 import options from '@/common/options';
-import optionsDefaults from '@/common/options-defaults';
-import hookSetting from '@/common/hook-setting';
+import { kScriptTemplate, kUpdateEnabledScriptsOnly } from '@/common/options-defaults';
 import { keyboardService } from '@/common/keyboard';
-import { focusMe, getActiveElement } from '@/common/ui';
+import { EXTERNAL_LINK_PROPS, focusMe, getActiveElement } from '@/common/ui';
+import { hookSettingsForUI } from '@/common/ui/util';
+import { store } from '@/options/utils';
+
+const items = {
+  autoUpdate: value => Math.max(0, Math.min(365, +value || 0)),
+  defaultInjectInto: { ...KNOWN_INJECT_INTO },
+  showAdvanced: value => value,
+  uiTheme: {
+    '': i18n('optionUiThemeAuto'),
+    dark: i18n('optionUiThemeDark'),
+    light: i18n('optionUiThemeLight'),
+  },
+  xhrInject: value => value,
+};
+</script>
+
+<script setup>
+import { onActivated, onDeactivated, reactive, ref, watch } from 'vue';
+import Tooltip from 'vueleton/lib/tooltip';
+import SettingCheck from '@/common/ui/setting-check';
 import LocaleGroup from '@/common/ui/locale-group';
 import SettingText from '@/common/ui/setting-text';
+import SettingsPopup from '@/common/ui/settings-popup.vue';
 import VmImport from './vm-import';
 import VmExport from './vm-export';
 import VmMaintenance from './vm-maintenance';
@@ -181,141 +146,30 @@ import VmEditor from './vm-editor';
 import VmBlacklist from './vm-blacklist';
 import VmDateInfo from './vm-date-info';
 
-const badgeColorEnum = {
-  badgeColor: i18n('titleBadgeColor'),
-  badgeColorBlocked: i18n('titleBadgeColorBlocked'),
-};
-const badgeColorNames = Object.keys(badgeColorEnum);
-const badgeColorItem = {
-  enum: badgeColorEnum, // exposing to the template
-  normalize: (value, name) => (
-    /^#[0-9a-f]{6}$/i.test(value) ? value : optionsDefaults[name]
-  ),
-};
-const items = {
-  autoUpdate: {
-    normalize: value => Math.max(0, Math.min(365, +value || 0)),
-  },
-  defaultInjectInto: {
-    enum: KNOWN_INJECT_INTO,
-  },
-  showAdvanced: {
-    normalize: value => value,
-  },
-  showBadge: {
-    enum: {
-      '': i18n('labelBadgeNone'),
-      unique: i18n('labelBadgeUnique'),
-      total: i18n('labelBadgeTotal'),
-    },
-  },
-  'filtersPopup.hideDisabled': {
-    enum: {
-      '': i18n('optionPopupShowDisabled'),
-      group: i18n('optionPopupGroupDisabled'),
-      hide: i18n('optionPopupHideDisabled'),
-    },
-  },
-  'filtersPopup.sort': {
-    enum: {
-      exec: i18n('filterExecutionOrder'),
-      alpha: i18n('filterAlphabeticalOrder'),
-    },
-  },
-  uiTheme: {
-    enum: {
-      '': i18n('optionUiThemeAuto'),
-      dark: i18n('optionUiThemeDark'),
-      light: i18n('optionUiThemeLight'),
-    },
-  },
-  xhrInject: {
-    normalize: value => value,
-  },
-  ...badgeColorEnum::mapEntry(() => badgeColorItem),
-};
-const normalizeEnum = (value, name) => (
-  hasOwnProperty(items[name].enum, value)
-    ? value
-    : Object.keys(items[name].enum)[0]
-);
-const getItemUpdater = (name, normalize) => (
-  debounce((value, oldValue) => {
-    value = normalize(value, name);
-    oldValue = normalize(oldValue, name);
-    if (value !== oldValue) options.set(name, value);
-  }, 50)
-);
+const $el = ref();
 const settings = reactive({});
+const expose = ref();
+const ctrlS = () => getActiveElement().dispatchEvent(new Event('ctrl-s'));
+let revokers;
 
-export default {
-  components: {
-    VmImport,
-    VmExport,
-    VmMaintenance,
-    VmSync,
-    VmEditor,
-    VmBlacklist,
-    VmDateInfo,
-    SettingCheck,
-    SettingText,
-    LocaleGroup,
-    Tooltip,
-  },
-  data() {
-    return {
-      expose: null,
-      items,
-      settings,
-    };
-  },
-  computed: {
-    editorWindowHint() {
-      return chrome.windows?.onBoundsChanged ? null : this.i18n('optionEditorWindowHint');
-    },
-    isCustomBadgeColor() {
-      return badgeColorNames.some(name => settings[name] !== optionsDefaults[name]);
-    },
-  },
-  methods: {
-    ctrlS() {
-      getActiveElement().dispatchEvent(new Event('ctrl-s'));
-    },
-    onResetBadgeColors() {
-      badgeColorNames.forEach(name => {
-        settings[name] = optionsDefaults[name];
-      });
-    },
-  },
-  activated() {
-    focusMe(this.$el);
-    this.revokers = [
-      keyboardService.register('ctrlcmd-s', this.ctrlS, { condition: 'inputFocus' }),
-    ];
-    items::forEachEntry(([name, { normalize = normalizeEnum }]) => {
-      this.revokers.push(hookSetting(name, val => { settings[name] = normalize(val, name); }));
-      this.$watch(() => settings[name], getItemUpdater(name, normalize));
-    });
-    this.expose = Object.keys(options.get(EXPOSE)).map(k => [k, decodeURIComponent(k)]);
-  },
-  deactivated() {
-    this.revokers.forEach((revoke) => { revoke(); });
-  },
-};
+onActivated(() => {
+  focusMe($el.value);
+  revokers = [
+    keyboardService.register('ctrlcmd-s', ctrlS, { condition: 'inputFocus' }),
+    ...hookSettingsForUI(items, settings, watch, 50),
+  ];
+  expose.value = Object.keys(options.get(EXPOSE)).map(k => [k, decodeURIComponent(k)]);
+});
+
+onDeactivated(() => {
+  revokers.forEach(r => r());
+  revokers = null;
+});
 </script>
 
 <style>
 .tab-settings {
   overflow-y: auto;
-  label {
-    display: inline-block;
-    > * {
-      vertical-align: middle;
-    }
-    &.valign-tb * {
-      vertical-align: text-bottom;
-    }
-  }
   input[type="number"] {
     width: 3.5em;
     padding-left: .25em;
